+++ scipy/linalg/src/_linalg_solve.hh
@@ -5,7 +5,7 @@
 #include <iostream>
 #include "numpy/arrayobject.h"
 #include "numpy/npy_math.h"
-#include "_ilp64_defines.h"
+#include "scipy_blas_defines.h"
 #include "_npymath.hh"
 #include "_common_array_utils.hh"
